--- todo_app/api/urls.py
@@ -1,13 +1,0 @@
-from django.urls import path
-
-from .views import (
-    UpdateTaskStatus,
-)
-
-urlpatterns = [
-    path(
-        "update-task-status/<int:task_id>/",
-        UpdateTaskStatus.as_view(),
-        name="update_task_status",
-    ),
-]
